sprint 12 commit 2

+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -5,4 +5,4 @@ class Track (
     val artistName: String,
     val trackTimeMillis: Long,
     val artworkUrl100: String,
-    val trackId: Double)
+    val trackId: Long)
